requirements.txt
# Web framework
Flask==3.1.0

# Authentication
Flask-Login==0.6.3

# Database - PostgreSQL (ORM)
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10

# Database - MongoDB
pymongo==4.10.1

# Session management
Flask-Session==0.8.0

# Environment variables
python-dotenv==1.0.1

# Password hashing & file utilities
Werkzeug==3.1.3

# Email sending
Flask-Mail==0.10.0

# BSON (MongoDB ObjectId)
bson==0.5.10

requests==2.32.3